[TM_WEB-4] Mark task as done
